Renamed OGFILE3.cpp to OGFILE_ARRAYS.cpp; merged OGFILE_REBELS.cpp back into it.
